+++ docs/.gitignore
@@ -1,0 +1,4 @@
+build/
+site/
+generated/
+examples/
