+++ docs/src/reference/advanced/graphquery.md
@@ -1,0 +1,32 @@
+# Advanced Graph Functions
+
+Some more advanced functions for quering a graph.
+
+Imported to namespace by
+```julia
+using NaiveNASlib.Advanced
+```
+
+```@docs
+findterminating
+```
+
+```@docs
+all_in_graph
+```
+
+```@docs
+output!
+```
+
+```@docs
+ancestors
+```
+
+```@docs
+descendants
+```
+
+```@docs
+defaultutility
+```
