+++ docs/src/reference/advanced/structure.md
@@ -1,0 +1,59 @@
+# Vertex Connection Strategies
+
+Strategies when changing connection between vertices. 
+
+Imported to namespace by
+```julia
+using NaiveNASlib.Advanced
+```
+```@docs
+RemoveStrategy
+```
+
+```@docs
+IncreaseSmaller
+```
+
+```@docs
+DecreaseBigger
+```
+
+```@docs
+AlignSizeBoth
+```
+
+```@docs
+PostAlign
+```
+
+```@docs
+ChangeNinOfOutputs
+```
+
+```@docs
+FailAlignSizeNoOp
+```
+
+```@docs
+FailAlignSizeError
+```
+
+```@docs
+FailAlignSizeWarn
+```
+
+```@docs
+NoSizeChange
+```
+
+```@docs
+CheckAligned
+```
+
+```@docs
+CheckNoSizeCycle
+```
+
+```@docs
+CheckCreateEdgeNoSizeCycle
+```
