+++ docs/src/reference/advanced/traits.md
@@ -1,0 +1,27 @@
+# Extra Vertex Traits
+
+Imported to namespace by
+```julia
+using NaiveNASlib.Advanced
+```
+
+
+```@docs
+NamedTrait
+```
+
+```@docs
+AfterΔSizeTrait
+```
+
+```@docs
+named
+```
+
+```@docs
+logged
+```
+
+```@docs
+validated
+```
