+++ docs/src/reference/extend/misc.md
@@ -1,0 +1,16 @@
+# Other Functions
+
+Imported to namespace by
+```julia
+using NaiveNASlib.Extend
+```
+
+```@docs
+parselect
+```
+```@docs
+vertex
+```
+```@docs
+NaiveNASlib.compconstraint!
+```
