+++ docs/src/reference/extend/strategies.md
@@ -1,0 +1,31 @@
+# Strategy Base Types 
+
+Useful when implementing new strategies.
+
+Imported to namespace by
+```julia
+using NaiveNASlib.Extend
+```
+
+```@docs
+AbstractΔSizeStrategy
+```
+
+```@docs
+AbstractJuMPΔSizeStrategy
+```
+
+```@docs
+DecoratingJuMPΔSizeStrategy
+```
+
+```@docs
+AbstractAfterΔSizeStrategy
+```
+```@docs
+AbstractAlignSizeStrategy
+```
+
+```@docs
+AbstractConnectStrategy
+```
