+++ docs/src/reference/extend/traits.md
@@ -1,0 +1,39 @@
+# Trait Types
+
+Traits are very useful for reasoning about a vertex without having to look at the wrapped computation.
+
+Imported to namespace by
+```julia
+using NaiveNASlib.Extend
+```
+
+```@docs
+trait
+```
+```@docs
+base(::DecoratingTrait)
+```
+```@docs
+MutationTrait
+```
+```@docs
+DecoratingTrait
+```
+```@docs
+Immutable
+```
+```@docs
+MutationSizeTrait
+```
+```@docs
+SizeAbsorb
+```
+```@docs
+SizeTransparent
+```
+```@docs
+SizeStack
+```
+```@docs
+SizeInvariant
+```
