+++ docs/src/reference/extend/vertices.md
@@ -1,0 +1,30 @@
+# Vertex Types
+
+While it is not generally expected that implementations should need to define new vertex types, the existing ones are typically useful for dispatching.
+
+Imported to namespace by
+```julia
+using NaiveNASlib.Extend
+```
+
+```@docs
+base(::AbstractVertex)
+```
+```@docs
+AbstractVertex
+```
+```@docs
+InputVertex
+```
+```@docs
+InputSizeVertex
+```
+```@docs
+CompVertex
+```
+```@docs
+MutationVertex
+```
+```@docs
+OutputsVertex
+```
