+++ docs/src/reference/simple/createvertex.md
@@ -1,0 +1,37 @@
+# Vertex Creation
+
+```@docs
+inputvertex
+```
+
+```@docs
+immutablevertex
+```
+
+```@docs
+absorbvertex
+```
+
+```@docs
+invariantvertex
+```
+
+```@docs
+conc
+```
+
+```@docs
++
+```
+
+```@docs
+-
+```
+
+```@docs
+*
+```
+
+```@docs
+/
+```
