+++ docs/src/reference/simple/graph.md
@@ -1,0 +1,26 @@
+# Graph Operations
+
+```@docs
+CompGraph
+```
+
+```@docs
+inputs(::CompGraph)
+```
+
+```@docs
+outputs(::CompGraph)
+```
+
+
+```@docs
+vertices
+```
+
+```@docs
+nvertices
+```
+
+```@docs
+findvertices
+```
