+++ docs/src/reference/simple/mutatevertex.md
@@ -1,0 +1,35 @@
+# Vertex Mutation
+
+```@docs
+Δnin!
+```
+
+```@docs
+Δnout!
+```
+
+```@docs
+Δsize!
+```
+
+```@docs
+insert!
+```
+
+```@docs
+remove!
+```
+
+```@docs
+create_edge!
+```
+
+```@docs
+remove_edge!
+```
+
+
+
+
+
+
