+++ docs/src/reference/simple/queryvertex.md
@@ -1,0 +1,17 @@
+# Access vertex data
+
+```@docs
+inputs(::AbstractVertex)
+```
+
+```@docs
+outputs(::AbstractVertex)
+```
+
+```@docs
+nin
+```
+
+```@docs
+nout
+```
